requirements.txt
astroid==3.3.5
bandit==1.7.10
black==24.10.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
curl_cffi==0.7.3
dill==0.3.9
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
idna==3.10
isort==5.13.2
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pbr==6.1.0
platformdirs==4.3.6
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
pylint==3.3.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.2
rsa==4.9
stevedore==5.3.0
tomli==2.0.2
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.3
